# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Education System API",
    "version": "1.0.0",
    "description": "Backend API for the Education System project. All responses use the wrapper `{ success, data, error }`. Success: `success: true`, `data: <payload>`, `error: null`. Failure: `success: false`, `data: null`, `error: { code, message, details }`. Every route is served under both /api and /api/v1 for one release. List endpoints share one pagination envelope { items, pagination: { page, limit, total, totalPages } } with limit max 50 and default 10. Common error codes: VALIDATION_ERROR (400), INVALID_JSON (400), UNSUPPORTED_FILE_TYPE (400), FILE_REQUIRED (400), UNAUTHORIZED (401), INVALID_CREDENTIALS (401), INVALID_CURRENT_PASSWORD (401), INVALID_RESET_TOKEN (401), INVALID_VERIFICATION_TOKEN (401), REFRESH_TOKEN_EXPIRED (401), REFRESH_TOKEN_REVOKED (401), INVALID_WEBHOOK_SIGNATURE (401), FORBIDDEN (403), SUBSCRIPTION_REQUIRED (403), NO_ACTIVE_SUBSCRIPTION (403), EMAIL_NOT_VERIFIED (403), TEACHER_NOT_FOUND (404), SUBSCRIPTION_NOT_FOUND (404), NOT_FOUND (404), SECTION_NOT_FOUND (404), ACTIVE_SUBSCRIPTION_EXISTS (409), SUBSCRIPTION_NOT_ACTIVE (409), ACTIVE_SUBSCRIPTIONS_EXIST (409), DUPLICATE_FIELD (409), AMBIGUOUS_USERNAME (409), FILE_TOO_LARGE (413), RATE_LIMIT_EXCEEDED (429). ACTIVE_SUBSCRIPTION_EXISTS means the calling student already holds one; ACTIVE_SUBSCRIPTIONS_EXIST means an admin delete is blocked by subscribers."
  },
  "servers": [
...
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/files/{name}": {
      "get": {
        "tags": ["Content"],
        "summary": "Download an uploaded file by generated name",
        "description": "Authorization matches the owning content gates: admins allow, owning teachers allow, students require an active subscription to the owning teacher's published content. Pass contentId to authorize against one specific content row; without it every row referencing the file must belong to covered teachers.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "contentId", "in": "query", "required": false, "schema": { "type": "integer" }, "description": "Authorize against one specific content row" },
          { "name": "download", "in": "query", "required": false, "schema": { "type": "string", "enum": ["1"] }, "description": "Set to 1 for Content-Disposition attachment instead of inline" }
        ],
        "responses": {
          "200": { "description": "File bytes with the stored content type" },
          "401": { "description": "Authentication required" },
          "403": { "description": "Access denied" },
          "404": { "description": "File not found" }
        }
      }
    },
...
  }
}